Move enum tests to the correct file
